[REFACTOR] 패키지 이름 변경
service 패키지에서 application 패키지로 이름 변경
