+++ src/main/java/techcourse/myblog/application/dto/ArticleDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.dto;
+package techcourse.myblog.application.dto;
 
 import techcourse.myblog.support.validator.TitleConstraint;
