+++ src/main/java/techcourse/myblog/application/dto/UserRequest.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.dto;
+package techcourse.myblog.application.dto;
 
 import techcourse.myblog.support.validator.EmailConstraint;
