+++ src/main/java/techcourse/myblog/application/exception/EditException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.exception;
+package techcourse.myblog.application.exception;
 
 public class EditException extends RuntimeException {
     public EditException(String message) {
