+++ src/main/java/techcourse/myblog/application/exception/ErrorMessage.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.exception;
+package techcourse.myblog.application.exception;
 
 public class ErrorMessage {
     private String errorMessage;
