+++ src/main/java/techcourse/myblog/application/exception/LoginException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.exception;
+package techcourse.myblog.application.exception;
 
 public class LoginException extends RuntimeException {
     public LoginException(String message) {
