+++ src/main/java/techcourse/myblog/application/exception/NotSameAuthorException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.exception;
+package techcourse.myblog.application.exception;
 
 public class NotSameAuthorException extends RuntimeException {
     public NotSameAuthorException(String message) {
